Move rest of controller logic
